# openapi3_0.yaml
openapi: 3.0.3
info:
  title: BIONIC
  description: |-
    
    Some useful links:
    - [Frontend repository.](https://github.com/kasia251011/car-app-frontend)
    - [Backend repository - server.](https://github.com/RosekKokosowy/BIONIC---server)
    - [Backend repository - microservice responsible for comparing cars.](https://github.com/RosekKokosowy/BIONIC---microservice_compare_cars)
    - [Backend repository - microservice responsible for selecting cars from database that are used during quiz.](https://github.com/RosekKokosowy/BIONIC---microservice_select_cars)
    - [Backend repository - microservice responsible for establishing priorities of the user based on the quiz.](https://github.com/RosekKokosowy/BIONIC---microservice_establish_priorities)
  termsOfService: http://swagger.io/terms/
  contact:
    email: 235847@edu.p.lodz.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
paths:
...
  /car/compare:
    get:
      tags:
        - comparison
      responses:
        '200':
          description: Successful operation
        '405':
          description: Invalid input
        '500':
          description: Internal error
    post:
      tags:
        - comparison
      summary: Requesting a comparison from server
      description: Requesting a comparison from a server about the objects user has provided.
      operationId: client-server-post-compare
      requestBody:
        description: Create a list of the cars to compare and the importance of its parameters.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarsInfo'
          application/xml:
            schema:
              $ref: '#/components/schemas/CarsInfo'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarId'
            application/xml:
              schema:
                $ref: '#/components/schemas/CarId'
        '405':
          description: Invalid input
        '500':
          description: Internal error
...
components:
  schemas:
    CarsInfo:
      type: object
      properties:
        carsParameters:
          type: array
          items:
            $ref: '#/components/schemas/CarParameters'
        parametersWeight:
          $ref: '#/components/schemas/ParametersWeight'
    CarParameters:
      type: object
      properties:
        id:
          type: integer
          format: int32
        yearOfManufacture:
          type: integer
          format: int32
        price:
          type: integer
          format: int64
        horsePower:
          type: integer
          format: int32
        typeOfFuel:
          type: string
        gearBox:
          type: string
        wheelDrive:
          type: string
      required:
        - id
        - yearOfManufacture
        - price
        - horsePower
        - typeOfFuel
        - gearBox
        - wheelDrive
...
    ParametersWeight:
      type: object
      properties:
        yearOfManufacture:
          type: number
          format: double
        mileage:
          type: number
          format: double
        price:
          type: number
          format: double
        horsePower:
          type: number
          format: double
        typeOfFuel:
          type: number
          format: double
        gearBox:
          type: number
          format: double
    CarId:
      type: object
      properties:
        id:
          type: integer
          format: int32
      required:
        - id
...